requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
duckdb==0.9.2
python-multipart==0.0.6
Werkzeug==3.0.1
python-dotenv==1.0.0
